fix some issues

+++ lib/models/result_place.dart
@@ -12,7 +12,7 @@ class ResultPlaceModel {
   double latitude;
   double longitude;
   List<String> types;
-  int userRatingsTotals;
+  double userRatingsTotals;
   GeneratedOptions generatedOptions;
 
   ResultPlaceModel(
